tests/wpt/web-platform-tests/workers/semantics/multiple-workers/005.html

<!--
if (location.hash == '#1') {
  var w2 = new SharedWorker('#2');
  w2.port.onmessage = function(e) {
    postMessage('1'+e.data);
  }
} else if (location.hash == '#2') {
  onconnect = function(e) {
    var port = e.ports[0];
    var w3 = new Worker('#3');
    w3.onmessage = function(e) {
      port.postMessage('2'+e.data);
    }
  }
} else {
  postMessage('3');
}

/*
-->
<!doctype html>
<title>dedicated worker in shared worker in dedicated worker</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
async_test(function() {
  var w1 = new Worker('#1');
  w1.onmessage = this.step_func(function(e) {
    assert_equals(e.data, '123');
    this.done();
  });
  w1.onerror = this.unreached_func("error");
});
</script>
<!--
*/
//-->
