LayoutTests/imported/w3c/web-platform-tests/workers/semantics/multiple-workers/005-expected.txt

CONSOLE MESSAGE: line 3: ReferenceError: Can't find variable: SharedWorker

Harness Error (FAIL), message = ReferenceError: Can't find variable: SharedWorker

TIMEOUT dedicated worker in shared worker in dedicated worker Test timed out
